+++ lang/tests/tests/named.rs
@@ -39,7 +39,7 @@ begin
   that
 
   let NamedInt : (item :: VType) =
-    (item = (NamedIdentity/constructor Int))
+    (item = (NamedIdentity/constructor Int64))
   that
 
   let ProjectedInt : VType = NamedInt/item that
@@ -78,7 +78,7 @@ begin
     end
   that
   let PunnedNamedInt : (Punned :: VType) =
-    PunnedNamedType Int
+    PunnedNamedType Int64
   that
   let PunnedProjectedInt : VType =
     PunnedNamedInt/Punned
@@ -86,7 +86,7 @@ begin
 
   let NestedNamedInt :
     (outer :: (inner :: VType)) =
-    (outer = (inner = Int))
+    (outer = (inner = Int64))
   that
   let NestedProjectedInt : VType =
     NestedNamedInt/inner
@@ -140,7 +140,7 @@ begin
       ret value
   } that
 
-  { ! named_identity (item = Int) 0 }
+  { ! named_identity (item = Int64) 0 }
 end
 "#,
     )
@@ -217,8 +217,8 @@ begin
     ) . Whole/item
   that
 
-  def payload_box : PayloadBox = (item = Int, 41) that
-  def whole_box : WholeBox = (item = Int, 42) that
+  def payload_box : PayloadBox = (item = Int64, 41) that
+  def whole_box : WholeBox = (item = Int64, 42) that
 
   let (item = A, payload) = payload_box in
   let (Whole, whole) = whole_box in
@@ -249,7 +249,7 @@ begin
     fn ((item = A, value) : Box) => ret value
   } that
 
-  def boxed : Box = (item = Int, 41) that
+  def boxed : Box = (item = Int64, 41) that
 
   { ! reveal boxed }
 end
@@ -263,7 +263,7 @@ fn rejects_named_term_with_mismatched_label() {
     NamedCase::assert_type_error(
         r#"
 begin
-  def bad : (x :: Int) = (y = 0) that
+  def bad : (x :: Int64) = (y = 0) that
   bad
 end
 "#,
@@ -275,7 +275,7 @@ fn rejects_named_pattern_with_mismatched_label() {
     NamedCase::assert_type_error(
         r#"
 begin
-  def value : (x :: Int) = (x = 0) that
+  def value : (x :: Int64) = (x = 0) that
   let (y = inner) = value in
   inner
 end
@@ -288,12 +288,12 @@ fn rejects_named_pattern_on_unnamed_mixed_component() {
     NamedCase::assert_type_error(
         r#"
 begin
-  let Mixed = (left :: Int) * (Int * (right :: Int)) that
+  let Mixed = (left :: Int64) * (Int64 * (right :: Int64)) that
   def value : Mixed = (left = 1, 2, right = 3) that
   let (
-    left = left : Int,
-    middle = middle : Int,
-    right = right : Int
+    left = left : Int64,
+    middle = middle : Int64,
+    right = right : Int64
   ) = value in
   (left, middle, right)
 end
@@ -306,11 +306,11 @@ fn rejects_mismatched_named_pattern_in_nested_mixed_product() {
     NamedCase::assert_type_error(
         r#"
 begin
-  let Nested = ((left :: Int) * Int) * (right :: Int) that
+  let Nested = ((left :: Int64) * Int64) * (right :: Int64) that
   def value : Nested = ((left = 1, 2), right = 3) that
   let (
-    (wrong = left : Int, middle : Int),
-    right = right : Int
+    (wrong = left : Int64, middle : Int64),
+    right = right : Int64
   ) = value in
   (left, middle, right)
 end
@@ -323,12 +323,12 @@ fn rejects_incompatible_named_payload_annotation_in_mixed_pattern() {
     NamedCase::assert_type_error(
         r#"
 begin
-  let Mixed = (left :: Int) * (Int * (right :: Int)) that
+  let Mixed = (left :: Int64) * (Int64 * (right :: Int64)) that
   def value : Mixed = (left = 1, 2, right = 3) that
   let (
     left = left : String,
-    middle : Int,
-    right = right : Int
+    middle : Int64,
+    right = right : Int64
   ) = value in
   (left, middle, right)
 end
@@ -370,7 +370,7 @@ fn rejects_missing_named_type_projection() {
         r#"
 begin
   let NamedInt : (item :: VType) =
-    (item = Int)
+    (item = Int64)
   that
   let InvalidProjection : VType =
     NamedInt/other
@@ -386,7 +386,7 @@ fn rejects_missing_named_projection() {
     NamedCase::assert_type_error(
         r#"
 begin
-  let Point = (x :: Int) * (y :: Int) that
+  let Point = (x :: Int64) * (y :: Int64) that
   def point : Point = (x = 0, y = 1) that
   point/z
 end
@@ -399,7 +399,7 @@ fn rejects_ambiguous_named_projection() {
     NamedCase::assert_type_error(
         r#"
 begin
-  let DuplicateFields = (x :: Int) * (x :: Int) that
+  let DuplicateFields = (x :: Int64) * (x :: Int64) that
   def duplicate : DuplicateFields = (x = 0, x = 1) that
   duplicate/x
 end
@@ -412,7 +412,7 @@ fn rejects_ambiguous_nested_named_projection() {
     NamedCase::assert_type_error(
         r#"
 begin
-  let DuplicateFields = ((x :: Int) * Int) * (outer :: (x :: Int)) that
+  let DuplicateFields = ((x :: Int64) * Int64) * (outer :: (x :: Int64)) that
   def duplicate : DuplicateFields = ((x = 0, 1), outer = x = 2) that
   duplicate/x
 end
@@ -426,7 +426,7 @@ fn rejects_ambiguous_nested_named_type_projection() {
         r#"
 begin
   let DuplicateFields : (x :: (x :: VType)) =
-    (x = (x = Int))
+    (x = (x = Int64))
   that
   let InvalidProjection : VType = DuplicateFields/x that
   ()
